+++ api/typescript/src/index.ts
@@ -4,7 +4,7 @@ export { KURTOSIS_VERSION } from "./kurtosis_version/kurtosis_version";
 // Services
 export type { FilesArtifactUUID, ContainerConfig } from "./core/lib/services/container_config";
 export { ContainerConfigBuilder } from "./core/lib/services/container_config";
-export type { ServiceID } from "./core/lib/services/service";
+export type { ServiceID, ServiceGUID } from "./core/lib/services/service";
 export { ServiceContext } from "./core/lib/services/service_context";
 export { PortSpec, PortProtocol } from "./core/lib/services/port_spec"